I can't totally decide how I want to handle authentication.
On the one hand, I don't want PAM and polkit to be forced on users who
don't want, need, or care about it.
On the other hand, authentication is super important and impacts the
package stack.

I also don't wan't to have multiple versions of packages just to
accomodate user choice - it makes dependency files difficult to parse or
properly enforce (it would lead us to the same spot, roughly).

This is the start to what I think is a 'clean' solution to the
authentication problem, and provides a minimal amount of friction. It
only exists at initial setup, and the choices are very easy to make (and
require minimal dependencies now). A user changing their mind would only
have to reinstall as many packages as they would if they decided to swap
device managers - but then again, why would they change their mind in
the first place? Authentication is a sort of set-it-and-forget it type
of thing; once you have your backend setup, you have no real need to
modify anything. Indeed, modifying things can cause breakages. Bad ones.

We'll see how this evolves, but I think this is a reasonable starting
point.